# Optimize SpatialBN on GPU