--- a/caffe2/operators/spatial_batch_norm_op.cc
+++ b/caffe2/operators/spatial_batch_norm_op.cc
@@ -6,6 +6,70 @@
 
 namespace caffe2 {
 
+template <>
+template <>
+void SpatialBNOp<CPUContext>::ComputeFusedParam<float>(
+    const int C,
+    const float* scale,
+    const float* bias,
+    const float* mean,
+    const float* var,
+    float* alpha,
+    float* beta) {
+  EigenVectorArrayMap<float> alpha_arr(alpha, C);
+  EigenVectorArrayMap<float> beta_arr(beta, C);
+  alpha_arr = ConstEigenVectorArrayMap<float>(scale, C) *
+      (ConstEigenVectorArrayMap<float>(var, C) + static_cast<float>(epsilon_))
+          .rsqrt();
+  beta_arr = ConstEigenVectorArrayMap<float>(bias, C) -
+      alpha_arr * ConstEigenVectorArrayMap<float>(mean, C);
+}
+
+template <>
+template <>
+void SpatialBNOp<CPUContext>::ComputeBatchMoments<float>(
+    const int N,
+    const int C,
+    const int HxW,
+    const float* batch_mean_sum,
+    const float* batch_var_sum,
+    float* mean,
+    float* var) {
+  const float scale = 1.0f / static_cast<float>(num_batches_ * N * HxW);
+  EigenVectorArrayMap<float> mean_arr(mean, C);
+  EigenVectorArrayMap<float> var_arr(var, C);
+  mean_arr = ConstEigenVectorArrayMap<float>(batch_mean_sum, C) * scale;
+  var_arr = ConstEigenVectorArrayMap<float>(batch_var_sum, C) * scale -
+      mean_arr.square();
+}
+
+template <>
+template <>
+void SpatialBNOp<CPUContext>::ComputeRunningMomentsAndFusedParam<float>(
+    const int C,
+    const float* scale,
+    const float* bias,
+    const float* mean,
+    const float* var,
+    float* running_mean,
+    float* running_var,
+    float* rstd,
+    float* alpha,
+    float* beta) {
+  const float a = 1.0f - momentum_;
+  const float b = momentum_;
+  math::Axpby<float, float, CPUContext>(C, a, mean, b, running_mean, &context_);
+  math::Axpby<float, float, CPUContext>(C, a, var, b, running_var, &context_);
+  math::InvStd<float, CPUContext>(
+      C, static_cast<float>(epsilon_), var, rstd, &context_);
+  EigenVectorArrayMap<float> alpha_arr(alpha, C);
+  EigenVectorArrayMap<float> beta_arr(beta, C);
+  alpha_arr = ConstEigenVectorArrayMap<float>(scale, C) *
+      ConstEigenVectorArrayMap<float>(rstd, C);
+  beta_arr = ConstEigenVectorArrayMap<float>(bias, C) -
+      alpha_arr * ConstEigenVectorArrayMap<float>(mean, C);
+}
+
 namespace {
 
 OpSchema::Cost CostInferenceForSpatialBN(