--- a/caffe2/operators/spatial_batch_norm_op.h
+++ b/caffe2/operators/spatial_batch_norm_op.h
@@ -204,15 +204,7 @@ class SpatialBNOp : public Operator<Context> {
       const T* mean,
       const T* var,
       T* alpha,
-      T* beta) {
-    EigenVectorArrayMap<T> alpha_arr(alpha, C);
-    EigenVectorArrayMap<T> beta_arr(beta, C);
-    alpha_arr = ConstEigenVectorArrayMap<T>(scale, C) *
-        (ConstEigenVectorArrayMap<T>(var, C) + static_cast<T>(epsilon_))
-            .rsqrt();
-    beta_arr = ConstEigenVectorArrayMap<T>(bias, C) -
-        alpha_arr * ConstEigenVectorArrayMap<T>(mean, C);
-  }
+      T* beta);
 
   template <typename T>
   void ComputeBatchMoments(
@@ -222,14 +214,7 @@ class SpatialBNOp : public Operator<Context> {
       const T* batch_mean_sum,
       const T* batch_var_sum,
       T* mean,
-      T* var) {
-    const T scale = T(1) / static_cast<T>(num_batches_ * N * HxW);
-    EigenVectorArrayMap<T> mean_arr(mean, C);
-    EigenVectorArrayMap<T> var_arr(var, C);
-    mean_arr = ConstEigenVectorArrayMap<T>(batch_mean_sum, C) * scale;
-    var_arr = ConstEigenVectorArrayMap<T>(batch_var_sum, C) * scale -
-        mean_arr.square();
-  }
+      T* var);
 
   template <typename T>
   void ComputeRunningMomentsAndFusedParam(
@@ -242,19 +227,7 @@ class SpatialBNOp : public Operator<Context> {
       T* running_var,
       T* rstd,
       T* alpha,
-      T* beta) {
-    const T a = T(1) - static_cast<T>(momentum_);
-    const T b = static_cast<T>(momentum_);
-    math::Axpby<T, T, Context>(C, a, mean, b, running_mean, &context_);
-    math::Axpby<T, T, Context>(C, a, var, b, running_var, &context_);
-    math::InvStd<T, Context>(C, static_cast<T>(epsilon_), var, rstd, &context_);
-    EigenVectorArrayMap<T> alpha_arr(alpha, C);
-    EigenVectorArrayMap<T> beta_arr(beta, C);
-    alpha_arr = ConstEigenVectorArrayMap<T>(scale, C) *
-        ConstEigenVectorArrayMap<T>(rstd, C);
-    beta_arr = ConstEigenVectorArrayMap<T>(bias, C) -
-        alpha_arr * ConstEigenVectorArrayMap<T>(mean, C);
-  }
+      T* beta);
 
   const bool is_test_;
   double epsilon_;
@@ -392,7 +365,8 @@ class SpatialBNGradientOp : public Operator<Context> {
           dbias_data,
           alpha_data,
           beta_data,
-          gamma_data);
+          gamma_data,
+          dX_data);
     }
     ComputeXGradient<T>(
         N, C, HxW, dY_data, X_data, alpha_data, beta_data, gamma_data, dX_data);
@@ -431,7 +405,8 @@ class SpatialBNGradientOp : public Operator<Context> {
       T* dbias,
       T* alpha,
       T* beta,
-      T* gamma);
+      T* gamma,
+      T* scratch);
 
   template <typename T>
   void ComputeXGradient(
@@ -452,6 +427,7 @@ class SpatialBNGradientOp : public Operator<Context> {
   Tensor alpha_;
   Tensor beta_;
   Tensor gamma_;
+  Tensor ones_;
 
   INPUT_TAGS(
       INPUT,