… caffe2/operators/spatial_batch_norm_op_gpu_impl.cuh