Summary: Optimize SpatialBN on GPU

Differential Revision: D13747581