@BIT-silence 2019-01-21
Benchmark

SpatialBNGradient for N = 128, C = 256, H = 128, W = 128,
NCHW order: 140.17ms -> 49.4ms (CuDNN 55.6ms)
NHWC order: 624.14ms -> 128.14ms.